pyproject.toml
[build-system]
build-backend = "hatchling.build"
requires = ["hatchling", "hatch-vcs"]

[project]
authors = [{name = "Remi Gau"}]
classifiers = [
    "Intended Audience :: Science/Research",
    "Intended Audience :: Developers",
    "License :: OSI Approved",
    "Programming Language :: C",
    "Programming Language :: Python",
    "Topic :: Software Development",
    "Topic :: Scientific/Engineering",
    "Operating System :: Microsoft :: Windows",
    "Operating System :: POSIX",
    "Operating System :: Unix",
    "Operating System :: MacOS",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12"
]
dependencies = [
    "pybids",
    "numpy",
    "nibabel"
]
description = "fake bids dataset generator"
# Version from setuptools_scm
dynamic = ["version"]
license = {text = "GNU GPL 3"}
maintainers = [{name = "Remi Gau", email = "remi.gau@gmail.com"}]
name = "fids"
readme = "README.md"
requires-python = ">=3.8"

[project.optional-dependencies]
# A combination of dependencies useful for developers
dev = [
    "fids[doc,test]",
    "pre-commit"
]
# Requirements necessary for building the documentation
doc = [
    "furo",
    "myst-parser",
    "numpydoc",
    "ruamel.yaml",
    "sphinx",
    "sphinx-copybutton"
]
# For running unit and docstring tests
test = [
    "coverage",
    "pytest",
    "pytest-cov"
]

[tool.black]
line-length = 89

[tool.codespell]
ignore-words-list = "nd"
skip = "./.git,p,.mypy_cache,env,venv,./doc/_build"

[tool.hatch.build.hooks.vcs]
version-file = "fids/_version.py"

[tool.hatch.build.targets.wheel]
packages = ["fids"]

[tool.hatch.version]
source = "vcs"

[tool.mypy]
check_untyped_defs = true
disallow_any_generics = true
disallow_incomplete_defs = true
disallow_untyped_defs = true
enable_error_code = ["ignore-without-code", "redundant-expr"]  # "truthy-bool"
no_implicit_optional = true
show_error_codes = true
# strict = true
warn_redundant_casts = true
warn_unreachable = true
warn_unused_ignores = true

[[tool.mypy.overrides]]
ignore_missing_imports = true
module = [
    "bids.*"
]

[[tool.mypy.overrides]]
ignore_errors = true
module = [
    'tests.*'
]

[tool.pytest.ini_options]
addopts = "-ra -q -vv --showlocals --strict-markers --strict-config"
# filterwarnings = ["error"]
# log_cli_level = "warning"
minversion = "6.0.0"
norecursedirs = "data"
testpaths = ["tests/"]
xfail_strict = true
